Fix conflicting functions and errors
originally, easye uses two different ints to grab the same thing: event.GetInt("victim").  we'll combine this into one for simplification and rename the variables as such. for the stats-concerning part of the Event_PassStolen function, we'll turn this into an if statement where it only runs if statsEnable.BoolValue returns TRUE.  then do the usual return Plugin_Handled.
we'll also change this check in Timer_DisplayStats for IsValidClient2 to 1. the only original difference between the two original functions was one line commented out that had the IsFakeClient check