remove dependency on compfixes
if we find the sm_projectiles_ignore_teammates command, set it to 0 so it doesn't do anything as we override it.
we save the current value from trikzProjCollide in trikzProjCollideSave (so whatever server.cfg sets it at)
we change the way dhooks works so when our value is changed it writes it down into an int value named trikzProjCollideCurVal. when the CbaseProjectile_cancollidewithteammates is called, we provide it with that current value and basically tell it how to work. now we have full control. also remove friendly fire check since we dont care
in trikz, we just change this value how we please and when we go back to 0, set it back to whatever it was before by using the saved value
fix up readme

this compiles and doesn't result in any on server start errors