trikz fix? also credit also new versioning system
basically we setup a separate hidden cvar that actually gets changed by the trikz function, and then we can override that with the publicly set cvar that stores an int that the player manually sets when they type in the cvar. by default it's two, but we can change it to one and it should work fine.
need to test tonight.
also credits to everyone for helping out :D. if this works like i think it should, 1.4.0 is around the corner.
plan is to version where:
Major.Minor.Patch
Major is huge functionality changes or huge code improvements
Minor is minor functionality changes or minor additions (like adding a new tracker)
Patch is just bugfixing