Change whitelist default and fix panaceas again
I tested it and if you exec pass_push, then ready up, it respawns players removing the offending items. so no biggie
also panaceas were still not working properly if the player landed before the goal was made so i changed it to where whenever they throw, if they're on the ground, panaceaCheck is false. works well